Organized a bit, placing Dashboard files in the appropriate folders.  We've decided to stick with static context for now, utilizing some hacks if a component needs multiple contexts.  Hilariously, tests worked on the first try with no additions.  App.js is now so clean we could probably use a functional component for it.  Couple of notes to ask about tomorrow:  1.)  Do we want to have state in each component, use the service directly from within each component?  I set up Dashboard to reference a function in context and store in the context state but either seems fine. 2.)  I'm using a global eslint config, but we had a config file floating around for it.  Should we not use global?  3.) Leading off of that, I'm going to standardize how I use ternary functions.  Should we do that for conditionals and ternaries all around?
