+++ src/App.js
@@ -1,20 +1,19 @@
 import React, { Component } from 'react';
-import Dashboard from './Dashboard';
-import config from './config'
+import Dashboard from './components/Dashboard/Dashboard';
 import ActivityForm from './components/ActivityForm/ActivityForm';
-import ActivityContext from './contexts/ActivityContext'
 
+/**
+ * We could probably change App to be a functional component -Blade 03-25-20
+ */
 class App extends Component {
 
   render() {
   return (
-    
     <main className='App'>
       <Dashboard />
       <ActivityForm />
     </main>
   );
-  
   }
 }
