+++ src/components/Dashboard/Dashboard.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import ActivityContext from './contexts/ActivityContext'
+import ActivityContext from '../../contexts/ActivityContext'
 
 export default class Dashboard extends Component {
   static contextType = ActivityContext;
@@ -48,8 +48,6 @@ export default class Dashboard extends Component {
       ? [Math.floor(Math.random() * this.context.activities.length)] 
       : 0;
     
-    console.log('inside the component:',this.context.activities)
-    
     return (
       <div>
         <h1>Indecisio</h1>
@@ -62,12 +60,12 @@ export default class Dashboard extends Component {
         </button>
         <div className="display-random-activity">
           <p>{this.context.activities && this.state.activityGenerated
-          ? `Your random activity is: ${this.context.activities[randomActivityIndex].name}`        
-          : ''}</p>
+            ? `Your random activity is: ${this.context.activities[randomActivityIndex].name}`        
+            : ''}</p>
           <p>
           {this.context.activities && this.state.activityGenerated
-          ? `The description is: ${this.context.activities[randomActivityIndex].description}`        
-          : ''}
+            ? `The description is: ${this.context.activities[randomActivityIndex].description}`        
+            : ''}
           </p>
         {this.state.activityGenerated && <button onClick={this.acceptRandomActivity}>Accept</button>}
         {this.state.activityGenerated && <button onClick={this.declineRandomActivity}>Decline</button>}
